requirements.txt
mcp23017==1.1.0
paho-mqtt==1.6.1
smbus==1.1.post2